Reduce size of congestion control next_*_event fields.
Since these are derived from the number of SENDMEs in a cwnd/cc update,
and a cwnd should not exceed ~10k, there's plenty of room in uint16_t
for them, even if the network gets significantly faster.

+++ src/core/or/congestion_control_st.h
@@ -158,11 +158,11 @@ struct congestion_control_t {
    * It is also reset to 0 immediately whenever the circuit's orconn is
    * blocked, and when a previously blocked orconn is unblocked.
    */
-  uint64_t next_cc_event;
+  uint16_t next_cc_event;
 
   /** Counts down until we process a cwnd worth of SENDME acks.
    * Used to track full cwnd status. */
-  uint64_t next_cwnd_event;
+  uint16_t next_cwnd_event;
 
   /** Are we in slow start? */
   bool in_slow_start;
